operator: operator/config: Introduce polymorphic advertised API config
The config in rpk was changed to support polymorphic advertised
kafka api. The latest redpanda container does not incorporate
mentioned change. To not create braking change for the user
operator introduces feature flag that will enable polymorphic
advertised kafka api in the configuration. The feature flag need
to be used after version v21.2.1 of the redpanda container.

When the feature flag is off, rpk's old config struct (copied here)
is used to generate the configuration file, otherwise we base
the configuration on rpk's new default that supports multiple socket
addresses.

Once the latest image is updated with this change, we need to
switch the default value of the flag, or remove it and the
corresponding code path entirely.