operator: operator/update: Handle cluster image changes for upgrade scenario
On a new redpanda image release we want the ability to upgrade the
cluster. This commit focuses on handling image changes in the redpanda
cluster CR by triggering a rolling update (using partitions) against
the statefulset underneath the CR.

The steps are as follows: 1) check statefulset image version identical
to the cluster CR; 2) if different, modify the image in the spec,
in-memory; 3) perform rolling update starting from the last pod to the
first; 4) in each iteration apply the update on the pod and requeue until
the pod is in ready state; 5) prior to a pod update ensure that the
previous pod has started listening to its Kafka API port - in another
we will ensure the restarted process has joined its replica groups.

A new status field, Upgrading, is also introduced. This allows us to
resume image updating after a requeue as well as controller termination
or crashing.