operator: operator/release: add make targets for pushing to dockerhub
New targets are added to the k8s project Makefile. These tag the
operator and configurator docker images based on the value of a
TAG_NAME variable, or with `:dev` tag whenever a commit is merged to the
main (dev) branch.

The second set of make targets are in charge of pushing the images to
dockerhub. In the case of images tagged with `TAG_NAME` (whenever there
is a new release), these push operations are idempotent. This is
accomplished by comparing the value contained in TAG_NAME with the
available tags on dockerhub, and only pushes when TAG_NAME is not
already on dockerhub. For `:dev` tags, no check is done as this tag is
intended to keep track to the tip of `dev`.