operator: operator: Add Prometheus ServiceMonitor to default manifest
Current definition was not working due to missing bearer token and TLS config
missing. Before bearer token was introduced, the Prometheus shows unathorized
error message when targets got scraped. When the bearer token was included in
ServiceMonitor configuration the Prometheus by default scrapes using the http
scheme. On the other hand the kube-rbac-proxy in current setup is using secure
listener which prevents from connection to be establish. After including the
HTTPS scheme the Promethus failed to scrape the Redpanda operator due to
server unknown certificate authority. If kube-rbac-proxy does not have provided
ca and proviate key, it will autogenerate mentioned certificates. End user can
configure certificate e.g. using cert-manager to include certificates ca
certificate in Prometheus to that can then validate server certificate. The
new kustomization with certificates need to be defined in volume and
volumeMounts. In the kube-rbac-proxy the `tls-cert-file` and
`tls-private-key-file` parameters need to be included in the args definition.

Prometheus that is included in kube-prometheus-stack helm chart will
autodiscover endpoints based on the service located under
`config/rbac/auth_proxy_service.yaml`.