operator: k8s/makefile: build image using multi-arch dockerfile
The Makefile is tweaked so that it obtains the platform in which `make`
is running (executes `uname -m`) and, using this value, creates a
`Dockerfile` with BUILDPLATFORM and TARGETARCH arguments that are later
passed via `--build-arg` flags. These values are used to instruct
`docker build` to create an image corresponding to the host platform
that can be used in e2e tests. This image is only used for testing; the
actual generation of multi-arch images is expected to be done
externally.

Signed-off-by: Ivo Jimenez <ivo@vectorized.io>