operator: go get -u github.com/spf13/cobra; task rpk:tidy
The next commit that introduces shell completion, which includes
completion for fish, requires an updated cobra.

The cobra we were using was highly out of date.

Updating the dependency updated a few other dependencies, but I'm
assuming that if the tests pass, things did not break -- looking at
cobra's release notes, there is no note of breakage, but the notes don't
even include the release that rpk was pinned to.