operator: go.{mod,sum}: task rpk:tidy
This pulls in two new deps, twmb/franz-go and twmb/tlscfg, which are
used in upcoming commits. The latter package is small, but is a
recreation and extension of something I created prior (in private). I've
found it to be the simplest way of initializing tls configurations I've
run into yet.