operator: k8s/tests: remove explicit kindest/node reference
Remove tag from kind config file so that the default kindest/node image
version is used. Kind hard-codes the SHA of the default image that has
been vetted by their community as the one that works well with the
version of kind being used. Using a specific tag is only needed if the
previous versions of the image need to be used.

Signed-off-by: Ivo Jimenez <ivo@vectorized.io>