operator: rpk topic describe -p: do not fail if ListOffsets fails
Describing partitions consists of two parts: the metadata half, and then
figuring out offsets. If we get to describing partitions (i.e. we do not
fail early), then we already have a successful metadata request. We
should print what we can and exit successfully if the listing fails.

It is common for partitions to have no leader when a broker restarts /
is down, or the partition is moving around. If a partition has no leader
yet, we can print a simple "-" for the offset, rather than failing the
request outright.

This pulls in a recent change in franz-go to collapse the per-partition
errors into one nice error message, and redoes some of PrintShardErr /
EachShard such that error messages are nicer.

The new output looks like so,

```
(ListOffsets failure: 1 partition has no leader from a metadata lookup)
PARTITION  LEADER  REPLICAS  LOG-START-OFFSET  HIGH-WATERMARK
0          0       [0]       -                 -
```