operator: acl user: redo, switch to admin.NewClient, add --mechanism
User creation and deletion previously used flags prefixed with "new-"
and "delete-" because the thought was we may need to have sasl flags to
even issue these requests. We do not need sasl to talk to the admin
handler, so we may as well drop "new-" and "delete-".

By the time Redpanda adds support for AlterUserScramCredentials, rpk
will have switched over to a global -X flag which will be used to
configure sasl.

We also add --mechanism to user creation, which allows specifying
SCRAM-SHA-512 (and also closes #2403). Supporting this requires passing
the mechanism to admin.CreateUser, so there are a few small
modifications there.

Tests have been removed due to two reasons:
1) the code remaining code is pretty tiny
2) coincidentally, the api/admin.go code tests user listing, creating,
and deleting.

So the only thing untested is the log line once the api call completes.