operator: k8s/types: Reserve 10% of memory for the OS.
If Redpanda runs inside a cgroup with a memory limit, then it should
not be allowed to consume all of the memory by default. There are
memory allocations and book-keeping done outside of Redpanda.

Setting `Resources.Requests.Memory` now reduces `Resources.RedpandaMemory()`
by 10% to allow for memory allocations outside of Redpanda.

If `Resources.Redpanda.Memory` is set, it will be honoured.

Redpanda is allowed to drop to 1.8Gi/core for backwards compatibility.