operator: rpk: add `debug logdirs describe`
This command adds support for the DescribeLogDirs request.

It appears that the IsFuture field is hardcoded to false, so we ignore
that column, and the OffsetLag is HWM - dirtyOffset, which I'm not sure
is correct and may potentially always be 1. Until we have clarity there,
we also leave out that column.

This also fixes API breakage from bumping kadm (which should soon be
stable), and now has `rpk debug bundle` on mac environment (with an
error), and reorganizes the debug.go file a tiny bit. We can do similar
cleanup as we address other commands.

Lastly, we fixup `rpk group seek --topics` to be a string slice flag
rather than string array flag.