operator: rpk: introduce --mode container
This commit introduce --mode container for
redpanda start, it's a set of well-known
cluster properties and flags for dev/container
environments

We are now in the path of deprecating
redpanda.developer_mode, so if its enabled
we will default to '--mode container'.

Container mode include overprovisioned, thats
why we are also adding overprovisioned to the
config.Default() method.