operator: rpk,k8s: bump franz-go
Recent updates in franz-go were made to fix some issues in rpk (and
allow fixing other issues).

Just by bumping this, we pull in new behavior where if we seek to the
future, we actually seek to the current end offsets (rather than -1).

This also returns a better error for when the controller is not ready:
"broker replied that the controller broker is not available"

Closes #6234.
Closes #6551.