operator: k8s: Fix console deletion validation
Console customer resource controller uses finalizers to stop deletion.
The main reason for finalizers is that controller is creating resources
outside of the k8s api. First resource is ACL and second is service account.
In order to correctly remove those resource the cluster reference, which is
part of the console spec, is used to build Admin API and Kafka client. Once
those resources are removed the finalizers should be removed. The problem
with such scenario is that finzalizers are removed by invoking client.Update
method. On the k8s api server side it triggeres an webhook for console. Due
to the fact that user trigger deletion webhook shouldn't check anything and
should pass immediately.